# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Bus Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: busId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Buses ──────────────────────────────────────────────────────────────────
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      parameters:
        - name: busId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed]
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Start a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip (e.g. end it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/scans:
    get:
      operationId: getTripScans
      tags: [trips]
      summary: Get all scans for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scan events for the trip
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanEvent"

  # ── Scans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, grade, guardianName, guardianPhone, biometricId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: string
        guardianName:
          type: string
        guardianPhone:
          type: string
        biometricId:
          type: string
        busId:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, grade, guardianName, guardianPhone, biometricId]
      properties:
        name:
          type: string
          minLength: 1
        grade:
          type: string
          minLength: 1
        guardianName:
          type: string
          minLength: 1
        guardianPhone:
          type: string
          minLength: 1
        biometricId:
          type: string
          minLength: 1
        busId:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        grade:
          type: string
        guardianName:
          type: string
        guardianPhone:
          type: string
        biometricId:
          type: string
        busId:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Bus ──────────────────────────────────────────────────────────────────
...
    Trip:
      type: object
      required: [id, busId, status, startedAt, createdAt]
      properties:
        id:
          type: integer
        busId:
          type: integer
        routeId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, completed]
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        totalBoardings:
          type: integer
        totalAlightings:
          type: integer
        createdAt:
          type: string
    TripInput:
      type: object
      required: [busId]
      properties:
        busId:
          type: integer
        routeId:
          type: ["integer", "null"]
    TripUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed]
        endedAt:
          type: ["string", "null"]

    # ── ScanEvent ─────────────────────────────────────────────────────────────
    ScanEvent:
      type: object
      required: [id, studentId, tripId, scanType, scannedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        tripId:
          type: integer
        scanType:
          type: string
          enum: [board, alight]
        scannedAt:
          type: string
        location:
          type: ["string", "null"]
        smsSent:
          type: boolean
        studentName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        busNumber:
          type: ["string", "null"]
...