# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Bus Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  grade: string;
  guardianName: string;
  guardianPhone: string;
  biometricId: string;
  /** @nullable */
  busId?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}