# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Bus Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  grade: string;
  /** @minLength 1 */
  guardianName: string;
  /** @minLength 1 */
  guardianPhone: string;
  /** @minLength 1 */
  biometricId: string;
  /** @nullable */
  busId?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  isActive?: boolean;
}